Update schema for integration
This commit concerns the integration of microservices and the need for databases to be consistent and to complement each other #7. In order to make bookings by querying the map service, it needs to contain postgres IDs for dental clinics, not mongodb objects, but also it needs clinic emails so that notification service can send email notifications.